[Feature-11195][UI] Add re-upload feature for resource files and udf files * fix resource file incorrect button link * add re-upload feature for resource files * add re-upload feature for udf files * modify udf edit button to rename button * fix file resource page button-link constraint * encapsulate file and udf resource code * place the breadcrumbs in the upper left corner * optimized some code * modify relative e2e test * Fix the problem of UDF page breadcrumbs when using the browser to go back or forward * adapt to the latest resource center
