+++ dolphinscheduler-e2e/dolphinscheduler-e2e-case/src/test/java/org/apache/dolphinscheduler/e2e/pages/resource/UdfManagePage.java
@@ -40,7 +40,7 @@ public class UdfManagePage extends NavBarPage implements ResourcePage.Tab {
     @FindBy(className = "btn-create-directory")
     private WebElement buttonCreateDirectory;
 
-    @FindBy(className = "btn-upload-udf")
+    @FindBy(className = "btn-upload-resource")
     private WebElement buttonUploadUdf;
 
     @FindBy(className = "items")
@@ -54,7 +54,7 @@ public class UdfManagePage extends NavBarPage implements ResourcePage.Tab {
 
     private final UploadFileBox uploadFileBox;
 
-    private final RenameDirectoryBox renameDirectoryBox;
+    private final RenameBox renameBox;
 
     private final CreateDirectoryBox createDirectoryBox;
 
@@ -63,7 +63,7 @@ public class UdfManagePage extends NavBarPage implements ResourcePage.Tab {
 
         uploadFileBox = new UploadFileBox();
 
-        renameDirectoryBox = new RenameDirectoryBox();
+        renameBox = new RenameBox();
 
         createDirectoryBox = new CreateDirectoryBox();
     }
@@ -106,15 +106,15 @@ public class UdfManagePage extends NavBarPage implements ResourcePage.Tab {
         udfList()
             .stream()
             .filter(it -> it.getText().contains(currentName))
-            .flatMap(it -> it.findElements(By.className("btn-edit")).stream())
+            .flatMap(it -> it.findElements(By.className("btn-rename")).stream())
             .filter(WebElement::isDisplayed)
             .findFirst()
             .orElseThrow(() -> new RuntimeException("No rename button in udf manage list"))
             .click();
 
-        renameDirectoryBox().inputName().clear();
-        renameDirectoryBox().inputName().sendKeys(AfterName);
-        renameDirectoryBox().buttonSubmit().click();
+        renameBox().inputName().clear();
+        renameBox().inputName().sendKeys(AfterName);
+        renameBox().buttonSubmit().click();
 
         return this;
     }
@@ -135,20 +135,20 @@ public class UdfManagePage extends NavBarPage implements ResourcePage.Tab {
     }
 
     @Getter
-    public class RenameDirectoryBox {
-        RenameDirectoryBox() {
+    public class RenameBox {
+        RenameBox() {
             PageFactory.initElements(driver, this);
         }
 
         @FindBys({
-                @FindBy(className = "input-directory-name"),
-                @FindBy(tagName = "input"),
+            @FindBy(className = "input-name"),
+            @FindBy(tagName = "input"),
         })
         private WebElement inputName;
 
         @FindBys({
-                @FindBy(className = "input-description"),
-                @FindBy(tagName = "textarea"),
+            @FindBy(className = "input-description"),
+            @FindBy(tagName = "textarea"),
         })
         private WebElement inputDescription;
