+++ dolphinscheduler-ui/src/locales/en_US/resource.ts
@@ -29,9 +29,11 @@ export default {
     file_name: 'File Name',
     description: 'Description',
     size: 'Size',
+    create_time: 'Create Time',
     update_time: 'Update Time',
     operation: 'Operation',
     edit: 'Edit',
+    reupload: 'ReUpload File',
     rename: 'Rename',
     download: 'Download',
     delete: 'Delete',
@@ -55,33 +57,9 @@ export default {
   },
   udf: {
     udf_resources: 'UDF resources',
-    create_folder: 'Create Folder',
     upload_udf_resources: 'Upload UDF Resources',
     udf_source_name: 'UDF Resource Name',
     user_name: 'Resource userName',
-    tenant_name: 'Resource tenantName',
-    whether_directory: 'Whether directory',
-    file_name: 'File Name',
-    file_size: 'File Size',
-    description: 'Description',
-    create_time: 'Create Time',
-    update_time: 'Update Time',
-    operation: 'Operation',
-    yes: 'Yes',
-    no: 'No',
-    edit: 'Edit',
-    rename: 'Rename',
-    download: 'Download',
-    delete: 'Delete',
-    delete_confirm: 'Delete?',
-    success: 'Success',
-    folder_name: 'Folder Name',
-    upload: 'Upload',
-    upload_files: 'Upload Files',
-    file_upload: 'File Upload',
-    enter_keyword_tips: 'Please enter keyword',
-    enter_name_tips: 'Please enter name',
-    enter_description_tips: 'Please enter description'
   },
   function: {
     udf_function: 'UDF Function',
