+++ dolphinscheduler-ui/src/views/resource/components/resource/index.module.scss
