+++ dolphinscheduler-ui/src/views/resource/components/resource/upload/index.tsx
@@ -14,18 +14,43 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { defineComponent, toRefs, PropType, getCurrentInstance } from 'vue'
+import { defineComponent, toRefs, PropType, getCurrentInstance, watch } from 'vue'
 import { NButton, NForm, NFormItem, NInput, NUpload } from 'naive-ui'
 
 import { useI18n } from 'vue-i18n'
 import Modal from '@/components/modal'
 import { useForm } from './use-form'
 import { useUpload } from './use-upload'
+import { ResourceType } from "@/views/resource/components/resource/types";
 
 const props = {
   show: {
     type: Boolean as PropType<boolean>,
     default: false
+  },
+  resourceType: {
+    type: String as PropType<ResourceType>,
+    default: undefined
+  },
+  isReupload: {
+    type: Boolean as PropType<boolean>,
+    default: false
+  },
+  name: {
+    type: String as PropType<string>,
+    default: ''
+  },
+  description: {
+    type: String as PropType<string>,
+    default: ''
+  },
+  fullName: {
+    type: String as PropType<string>,
+    default: ''
+  },
+  userName: {
+    type: String as PropType<string>,
+    default: ''
   }
 }
 
@@ -58,6 +83,21 @@ export default defineComponent({
     }
 
     const trim = getCurrentInstance()?.appContext.config.globalProperties.trim
+
+
+    watch(
+      () => props.show,
+      () => {
+        state.uploadForm.type = props.resourceType!
+        state.uploadForm.isReupload = props.isReupload
+        if (props.isReupload && props.show) {
+          state.uploadForm.fullName = props.fullName
+          state.uploadForm.name = props.name
+          state.uploadForm.description = props.description
+          state.uploadForm.user_name = props.userName
+        }
+      }
+    )
 
     return {
       hideModal,
