+++ dolphinscheduler-ui/src/views/resource/components/resource/upload/use-upload.ts
@@ -18,7 +18,7 @@
 import { useI18n } from 'vue-i18n'
 import { IEmit } from '../types'
 import { useFileStore } from '@/store/file/file'
-import { createResource } from '@/service/modules/resources'
+import { createResource, updateResource } from '@/service/modules/resources'
 
 export function useUpload(state: any) {
   const { t } = useI18n()
@@ -34,16 +34,24 @@ export function useUpload(state: any) {
     if (state.saving) return
     state.saving = true
     try {
-      // no more pid, as currentDir acts as the pid or parent path right now.
-      const currentDir = fileStore.getCurrentDir || '/'
       const formData = new FormData()
       formData.append('file', state.uploadForm.file)
-      formData.append('type', 'FILE')
+      formData.append('type', state.uploadForm.type)
       formData.append('name', state.uploadForm.name)
-      formData.append('currentDir', currentDir)
       formData.append('description', state.uploadForm.description)
 
-      await createResource(formData as any)
+      if (state.uploadForm.isReupload) {
+        formData.append('user_name', state.uploadForm.user_name)
+        formData.append('fullName', state.uploadForm.fullName)
+        formData.append('tenantCode', state.uploadForm.user_name)
+        await updateResource(formData as any)
+      } else {
+        // no more pid, as currentDir acts as the pid or parent path right now.
+        const currentDir = fileStore.getCurrentDir || '/'
+        formData.append('currentDir', currentDir)
+        await createResource(formData as any)
+      }
+
       window.$message.success(t('resource.file.success'))
       state.saving = false
       emit('updateList')
