+++ dolphinscheduler-ui/src/views/resource/components/resource/use-file.ts
@@ -21,12 +21,13 @@ import {
   viewResource
 } from '@/service/modules/resources'
 import type { ResourceListRes } from '@/service/modules/resources/types'
-import { IResourceListState, ISetPagination } from './types'
+import { IResourceListState, ISetPagination, ResourceType } from './types'
 
 export function useFileState(
   setPagination: ISetPagination = {} as ISetPagination
 ) {
   const getResourceListState: IResourceListState = (
+    type: ResourceType,
     fullName = '',
     tenantCode = '',
     searchVal = '',
@@ -37,7 +38,7 @@ export function useFileState(
       queryResourceListPaging({
         fullName,
         tenantCode,
-        type: 'FILE',
+        type: type,
         searchVal,
         pageNo,
         pageSize
@@ -56,6 +57,7 @@ export function useFileState(
             file_name: item.fileName,
             description: item.description,
             size: item.size,
+            create_time: item.createTime,
             update_time: item.updateTime
           }
         })
